% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_round.R
\name{base_round}
\alias{base_round}
\title{Basic and semi-controlled random rounding for tabulated data}
\usage{
base_round(
  obj,
  base = 5,
  controlled_ran = FALSE,
  fewer_than_base = paste0("fewer than ", base)
)
}
\arguments{
\item{obj}{an \code{aggregate_data} object}

\item{base}{numeric, can take any integer value greater than zero. Values in your data will be rounded to a
multiple of \code{base}.}

\item{controlled_ran}{logical. If \code{FALSE}, values in the data will simply be rounded to the
nearest multiple of \code{base}. Totals will not be preserved. If set to \code{TRUE} then semi-controlled random rounding will
be applied to the data in order to preserve the overall total. Marginal totals will not be
preserved. This introduces some distortion into your data. See 'Details' below for more information about semi-controlled random rounding.}

\item{fewer_than_base}{character or statement that yields a value with class character. A message to display
in cells that are rounded down to 0 to distinguish them from "true" 0 values where no observations
were recorded. See "Details" below for more information about best practice in representing rounded 0 values.}
}
\value{
A data frame with the same dimensions and format as \code{obj$original_data},
with rounding applied according to your choice of \code{base}.
}
\description{
\code{base_round} finds locations in your data where the
specified cell is less than or equal to a cutoff value of your choosing.
The function prints off a list of locations that can optionally be stored as variable.
}
\details{
Semi-controlled random rounding is carried out according to the method described in section 3.2
of the paper \href{https://sciendo.com/article/10.1515/jos-2015-0019}{Measuring Disclosure Risk and Data Utility for Flexible Table Generators} (Shlomo et al, 2015).
While "fully controlled" random rounding seeks to preserve the additivity of all marginal totals of a data table, semi-controlled
random rounding only preserves the (rounded) overall table total. This is to say that the overall total
given by \code{base_round} in the output data will be the same as if the total in the input data were rounded to the
nearest multiple of \code{base} and the rounded marginal totals for rows and columns will each add up to that overall total.
However, each column may not add up to its individual marginal total after applying semi-controlled random rounding.

This is perhaps easier to grasp when presented in the form of an example. The table \code{unrounded_data}
below has not has any rounding applied to it. Therefore, risk of disclosure may be present.

\verb{data frame name: unrounded_data}\tabular{ccccc}{
   \strong{treatment}     \tab \strong{data1}     \tab \strong{data2}     \tab \strong{data3}     \tab \strong{total_col}     \cr
   \strong{A} \tab 0 \tab 5 \tab 1 \tab 6 \cr
   \strong{B} \tab 4 \tab 2 \tab 2 \tab 8 \cr
   \strong{C} \tab 1 \tab 7 \tab 3 \tab 11 \cr
   \strong{D} \tab 2 \tab 2 \tab 2 \tab 6 \cr
   \strong{total_row} \tab 7 \tab 16 \tab 8 \tab 31 \cr
}


The table \code{simple_rounded} below has had all values rounded to the nearest multiple of three with no control applied,
equivalent to calling \code{base_round()} with \code{base = 3} and \code{controlled_ran = FALSE}. We do not remove the marginal
totals before rounding and recalculate them afterwards as this would introduce significant rounding error,
especially for very large tables, causing us to lose vital information.

All values in the table are rounded as closely to their original values as possible, but all of the original
values have been obscured, helping to reduce the risk of disclosure. However, the marginal totals no
longer align with the table values nor with the the overall total, suggesting that methods which prioritise closeness
to the original data  at the level of individual cells nevertheless incur some loss of information at a wider level.

\verb{data frame name: simple_rounded}\tabular{ccccc}{
   \strong{treatment}     \tab \strong{data1}     \tab \strong{data2}     \tab \strong{data3}     \tab \strong{total_col}     \cr
   \strong{A} \tab 0 \tab 6 \tab fewer than 3 \tab 6 \cr
   \strong{B} \tab 3 \tab 3 \tab 3 \tab 9 \cr
   \strong{C} \tab fewer than 3 \tab 6 \tab 3 \tab 12 \cr
   \strong{D} \tab 3 \tab 3 \tab 3 \tab 6 \cr
   \strong{total_row} \tab 6 \tab 15 \tab 9 \tab 30 \cr
}


The table \code{semi_controlled_rounding} below has had base-3 semi-controlled random rounding applied to it, equivalent to calling
\code{base_round()} with \code{base = 3} and \code{controlled_ran = TRUE}. We see that as before, all cells have been rounded up or down to
a multiple of three, but rather than round to the closest multiple, we have randomly chosen whether each
cell is rounded up or down.\tabular{ccccc}{
   \strong{treatment}     \tab \strong{data1}     \tab \strong{data2}     \tab \strong{data3}     \tab \strong{total_col}     \cr
   \strong{A} \tab 0 \tab 6 \tab fewer than 3 \tab 6 \cr
   \strong{B} \tab 3 \tab 3 \tab 3 \tab 9 \cr
   \strong{C} \tab 3 \tab 6 \tab 3 \tab 12 \cr
   \strong{D} \tab fewer than 3 \tab 3 \tab 3 \tab 6 \cr
   \strong{total_row} \tab 6 \tab 15 \tab 9 \tab 30 \cr
}
}
\examples{
To do
}
