% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregate_data.R
\name{create_aggregate_data}
\alias{create_aggregate_data}
\title{Create an aggregate_data object}
\usage{
create_aggregate_data(
  original_data,
  data,
  treatment,
  total_col = NULL,
  total_row = NULL
)
}
\arguments{
\item{original_data}{A data frame that you wish to convert to an \code{aggregate_data} object. See details for
more information on required format.}

\item{data}{A vector of indices indicating the columns in original_data that contain numeric data.}

\item{treatment}{A single numeric value indicating the index of the column in original_data that contains row names.}

\item{total_col}{Optional. A single numeric value indicating the index of the column in original_data that contains row totals.}

\item{total_row}{Optional. A single numeric value indicating the index of the row in original_data that contains column totals.}
}
\value{
An aggregate_data object
}
\description{
Convert a data frame into an object of class \code{aggregate_data}, the required format
for use with sdcUKHSA functions.
}
\details{
Aggregate data must be converted to an object of class \code{aggregate_data} before it can be used
with sdcUKHSA functions. Your data should be loaded into R as a data frame, at which point
it can be converted to an \code{aggregate_data} object using the \code{create_aggregate_data()} function.

\code{create_aggregate_data()} expects \code{original_data} to take the form, essentially, of a two-way
table, with rows indicating groups of one categorical variable and columns indicating groups
of another categorical variable. This requires your data to have certain features which must
be indicated in the function call:

\itemize{
\item One column specifying groups of a treatment or other categorical variable.
The index of this column is passed via the \code{treatment} parameter.
\item Columns specifying the numerical observations in your data, broken down into groups of some
how many observations for each group in \code{treatment}.
}

Data of this type often includes totals for each row and/or column. These can be helpful to users
and although this is not required
to create an aggregate_data object.
If your data contains marginal totals for the rows and/or columns of \code{data}, these must be
indicated as well:

\itemize{
\item If marginal \strong{column} totals are present, the \strong{row} containing these must be specified
by passing its index to the \code{total_row} parameter.
\item If marginal \strong{row} totals are present, the \strong{column} containing these must be specified
by passing its index to the \code{total_col} parameter.
}

The table below shows an example of data in an appropriate format:

\verb{data frame name: example_data}\tabular{ccccc}{
   \strong{treatment}     \tab \strong{data1}     \tab \strong{data2}     \tab \strong{data3}     \tab \strong{total_col}     \cr
   \strong{A} \tab 0 \tab 5 \tab 1 \tab 6 \cr
   \strong{B} \tab 4 \tab 2 \tab 2 \tab 8 \cr
   \strong{C} \tab 1 \tab 7 \tab 3 \tab 11 \cr
   \strong{total_row} \tab 5 \tab 14 \tab 6 \tab 25 \cr
}


The corresponding function call to convert this table into an \code{aggregate_data} object would be:
(recall that column names are not treated)
\verb{create_aggregate_data(original_data = example_data, data = c(2, 3, 4), treatment = 1, total_col = 5, total_row = 4}
}
\examples{
To do
}
