% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_small_values.R
\name{find_small_values}
\alias{find_small_values}
\title{Check for small values in tabular data}
\usage{
find_small_values(obj, cutoff = 5, bycol = TRUE)
}
\arguments{
\item{obj}{an \code{aggregate_data} object}

\item{cutoff}{numeric. Any entries less than or equal to \code{cutoff} will be included in the output.}

\item{bycol}{logical, indicating whether the output should be sorted by column (\code{TRUE}) or row (\code{FALSE}).}
}
\value{
A matrix with one row per value in the input data that is less than or equal to \code{cutoff}.
There are two columns in the matrix, \code{row} and \code{col}, which give the row and column number of each
flagged value, respectively.
}
\description{
\code{find_small_values} finds locations in your data where the
specified cell is less than or equal to a cutoff value of your choosing.
The function prints off a list of locations that can optionally be stored as variable.
}
\examples{
To do
}
