% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_zeros.R
\name{find_zeros}
\alias{find_zeros}
\title{Check for non-structural zeros in tabular data}
\usage{
find_zeros(obj, prop = 0.1)
}
\arguments{
\item{obj}{an \code{aggregate_data} object}

\item{prop}{numeric, must be between 0 and 1. Any columns with a
proportion of zeros that is greater than or equal to \code{prop} will be included in the output.}
}
\value{
A matrix with one row per column in the input data that has a proportion of zeros
greater than or equal to \code{prop}. There are two columns in the matrix, \code{col} and \code{proportion},
which give the column name and proportion of zeros in that column, respectively.
}
\description{
\code{find_zeros} finds columns in your data that have a high proportion of
non-structural zeros as defined by a user-chosen cutoff
}
\examples{
To do
}
