% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anonymity.R
\name{k_anonymity}
\alias{k_anonymity}
\title{Check for non-structural zeros in tabular data}
\usage{
k_anonymity(data, k)
}
\arguments{
\item{data}{a data frame of microdata with one observation per row}

\item{k}{numeric, combinations of characteristics that occur fewer than \code{k} times will be found}

\item{cols}{optional, a vector of indices indicating the columns in \code{data} that will be used
in generating combinations to search. By default, all columns in \code{data} are used}
}
\value{
a data frame. This will have one row for each unique combination of the values of \code{cols}
that occurs fewer than k times, and one column corresponding to each column in \code{cols}. The last
(rightmost) column in this data frame will be numeric and give the number of times that each combination occurs.
}
\description{
\code{k_anonymity} finds combinations of characteristics in microdata observations
that occur fewer than \code{k} times
}
\examples{
To do
}
